Creates an auth script to protect routes and open sign in dialog
Uses rootScope to prevent a state change from welcome if the user
does not have a token -- we don't have any public pages yet, so
by default everything is protected. If the user doesn't have a
token, then click go no where and a dialog box with sign in forms
opens. On success, dialog closes and user is routed to clicked link.