Sets up controller to be ready for LinkedIn data
There's nothing in the resolve block on the router (yet) to handle LI data, but once we get that working, we'll need to put everything into the $ctrl for the template to render. 

Assuming that the names in the LI api docs will match the json data that comes back -- might need some minor changes in variable names.