Renames component from profile to linkedIn, since it's li data only
The userData should be coming in through the routes via a parent
component's resolve block. We're pretending to pass a userData object
in and then pulling data off of it to put on the controller. Not
clear if the properties we're using are going to match the json data
we get back from the linkedIn API.