Adds a userUrl to model, public user routes, and user-service
We need a public route to get the data which will be rendered for
a user's profile. This creates the routes in the front end user-
service, which now calls an unauthenticated route in the backend
/users/:userUrl which searches for a user's data by their unique
url. Adds the routes to the Express router in user-router and in
the app.js