Updates styling for home page to have menu bar, boxes, sign in / out
The home component is now the default when you hit / instead of
being at /home. It has most of the styling in the component instead
of in the main, which will have to move when we move the menu bar
to index.html.

The sign in and out button toggles depending on if the user has a
token saved or not. Signing in takes a user to their profile page.