Moves all of the nav controls/styles out of home into a component
So the home and profile page can use the same nav bar, moved all of
it's styling into main, and then moved the controller and template
bits into a separate nav component.

The home and profile routes _dont_ have named views, instead, they
include the nav as a component directly.