Adds ngDialog success function to personal info component
The personal info component was calling a save method in the
userService, but not the dialog box's success method, which closes
the dialog. This fixes the dialog to close upon submit, and the
reset is done at the loading of the component, instead of on the
save button being clicked.