Standardizes LI data object to be consistent, enables updating btn components
The data for linked in needs to be able to be grabbed in the linkedIn
component, passed to the parent, and to the get-linked, which is nested
inside an ngDialog. The easiest way to do this was to make the parent
user-dash component the owner of the savedLink object, and to give the
children two-way bindings to grab it and update it.

In order to get the data through ngDialog, it has to be explicitly passed
into the controller. This is vital for getting everything through.

Works -- saving or updating data now updates the parent.