Adds ensureToken, ensureAdmin middleware to admin routes to start the authorization checks
The token now includes putting user's roles in an array on their
token. This roles array is checked for an 'admin' entry, if found
it passes back to the route, otherwise it throws a 401 error.

Using a plain-text array of roles is almost certainly a terrible
idea. This is just to get the structure up and running before
getting authorization working.