Updates server and app packages, adds a top level for CI & testing
To get travis CI to work, there needs to be a top level package.json
that tells it to jump into each side of the project, install
dependencies, and then run tests. Added this script.

Changed the licenses to all match as ICS, and fixed bad copied repo
links from an old project (whoops). Configs should more or less
match at this point.